requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
Flask-Mail
APScheduler
psycopg2-binary
python-dotenv
python-dateutil
XlsxWriter